MDEV-13577 slave_parallel_mode=optimistic should not report the mode's            specific temporary errors
The optimistic parallel slave's worker thread could face an error due to
the algorithm's specifics which allows for conflicts like the reported
"Can't find record in 'table'".
The conflicts are resolved with rollback and retry and therefore
do have to confuse an error log reader with their empty urgency.

The case is fixed with refining my_message_sql() to downgrade formerly
error level of an error message to optionally (when --log-warnings=2)
the warning one when the error comes from the optimistically run
parallel slave thread.

This downgrade decision does not leave any escape path for a permanent error
which continues to be reported by the slave driver thread
to all logging media, the error log incl.